pyproject.toml
[tool.poetry]
name = "multitenancy-fastapi"
version = "0.1.0"
description = ""
authors = ["TIZadid <talhazadid@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.3"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
uvicorn = "^0.32.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
